last updates on time period and configuration setings

+++ ForcePlatformData/Models/AppsettingsModel.cs
@@ -5,7 +5,7 @@
         public bool AutoSelectCom { get; set; }
         public string ComPort { get; set; }
         public int FilterLength { get; set; }
-        public double CalibrateZ { get; set; }
+        public int CalibrateZ { get; set; }
         public double FreeFallAcc { get; set; }
         public string ReportsPath { get; set; }
         public string ChromePath { get; set; }
